requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.4
dataclasses-json==0.6.7
Deprecated==1.3.1
distro==1.9.0
duckdb==1.4.1
faiss-cpu==1.12.0
frozenlist==1.8.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.11.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.17
langchain-community==0.2.19
langchain-core==0.2.43
langchain-experimental==0.0.42
langchain-openai==0.1.25
langchain-text-splitters==0.2.4
langgraph==0.2.20
langgraph-checkpoint==1.0.12
langsmith==0.1.147
marshmallow==3.26.1
msgpack==1.1.2
mss==10.1.0
multidict==6.7.0
mypy_extensions==1.1.0
nodriver==0.48.0
numpy==1.26.4
openai==1.109.1
orjson==3.11.4
packaging==24.2
propcache==0.4.1
pydantic==2.12.4
pydantic_core==2.41.5
pypdf==6.1.3
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.44
tenacity==8.5.0
tiktoken==0.12.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
websockets==15.0.1
wrapt==2.0.1
yarl==1.22.0
